refactor: nest verify command into src/verify/ + shared src/utils/ (#443)
Addresses review: don't dump the command's files flat into src/. The
`verify` command now lives in its own folder with the `npm-tarball`
subcommand in a subfolder, and the generic file-walk helper moves to a
shared utils folder:

  src/verify/
    posture.ts / posture-types.ts / posture-format.ts   (#414, moved)
    npm-tarball/
      index.ts main.ts triple.ts types.ts
      resolve-url.ts download.ts local-dir-state.ts
  src/utils/
    list-files-recursive.ts

Pure move + import-path updates; no behavior change. Extending the
one-folder-per-command layout to plan/publish/status/etc. is a separate
repo-wide refactor.
