+++ src/cli.ts
@@ -48,9 +48,9 @@ import { publish } from './publish.js';
 import { reconcile } from './reconcile.js';
 import { formatStatusRow } from './status-format.js';
 import { computeStatus } from './status.js';
-import { formatVerifyRow } from './verify-format.js';
-import { verifyNpmTarball } from './verify-npm-tarball.js';
-import { computeVerify } from './verify.js';
+import { formatVerifyRow } from './verify/posture-format.js';
+import { verifyNpmTarball } from './verify/npm-tarball/index.js';
+import { computeVerify } from './verify/posture.js';
 import { writeCrateVersionForBuild } from './write-crate-version.js';
 import { writeLauncherFromConfig } from './write-launcher.js';
 import { writeVersionForBuild } from './write-version.js';
