+++ src/verify/npm-tarball/local-dir-state.test.ts
