+++ src/verify/posture-format.ts
@@ -5,7 +5,7 @@
  * Issue #414.
  */
 
-import type { VerifyRow } from './verify-types.js';
+import type { VerifyRow } from './posture-types.js';
 
 export function formatVerifyRow(row: VerifyRow): string {
   const mark = row.posture === 'oidc' ? '✓' : row.posture === 'token' ? '⚠' : '?';
