+++ src/verify/posture.test.ts
@@ -15,7 +15,7 @@ import { tmpdir } from 'node:os';
 import { join } from 'node:path';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 
-import { run } from './cli.js';
+import { run } from '../cli.js';
 
 let repo: string;
 const stdoutChunks: string[] = [];
